requirements.txt
fastapi
uvicorn
pandas
networkx
python-dotenv
motor
pymongo
pydantic
